Fix long tile image loading

+++ OsmAnd/src/net/osmand/plus/views/layers/MapMarkersLayer.java
@@ -830,8 +830,8 @@ public class MapMarkersLayer extends OsmandMapLayer implements IContextMenuProvi
 
 		VectorLineBuilder outlineBuilder = new VectorLineBuilder();
 		VectorDouble outlinePattern = new VectorDouble();
-		outlinePattern.add(75 / getMapDensity());
-		outlinePattern.add(55 / getMapDensity());
+		outlinePattern.add(75.0d / (double) getMapDensity());
+		outlinePattern.add(55.0d / (double) getMapDensity());
 		FColorARGB outlineColor = new FColorARGB(1.0f, 1.0f, 1.0f, 1.0f);
 		double strokeWidth = 20.0d;
 		int outlineId = isLast ? 20 : 10;
